+++ lua/config/plugins/harpoon.lua
@@ -1,7 +1,33 @@
 return {
-	{
-		"ThePrimeagen/harpoon",
-		branch = "harpoon2",
-		dependencies = { "nvim-lua/plenary.nvim" },
-	},
+	"ThePrimeagen/harpoon",
+	branch = "harpoon2",
+	dependencies = { "nvim-lua/plenary.nvim" },
+	config = function()
+		local harpoon = require("harpoon")
+		-- REQUIRED
+		harpoon:setup()
+		-- REQUIRED
+
+		vim.keymap.set("n", "<leader>a", function()
+			harpoon:list():add()
+		end)
+
+		vim.keymap.set("n", "<C-e>", function()
+			harpoon.ui:toggle_quick_menu(harpoon:list())
+		end)
+
+		--maybe in the future
+		--vim.keymap.set("n", "<C-h>", function() harpoon:list():select(1) end)
+		--vim.keymap.set("n", "<C-t>", function() harpoon:list():select(2) end)
+		--vim.keymap.set("n", "<C-n>", function() harpoon:list():select(3) end)
+		--vim.keymap.set("n", "<C-s>", function() harpoon:list():select(4) end)
+
+		-- Toggle previous & next buffers stored within Harpoon list
+		vim.keymap.set("n", "<leader>hp", function()
+			harpoon:list():prev()
+		end)
+		vim.keymap.set("n", "<leader>hn", function()
+			harpoon:list():next()
+		end)
+	end,
 }
