--- lua/config/plugins/utilis.lua
@@ -1,8 +1,0 @@
-return {
-	"tpope/vim-fugitive",
-	"mbbill/undotree",
-	"tpope/vim-rhubarb", -- GitHub extension for fugitive (enables :GBrowse)
-	"nvimtools/none-ls.nvim", -- For formatting with prettier
-	"rafamadriz/friendly-snippets", -- snippets(not worked need to figure our why)
-	"windwp/nvim-autopairs", -- Automatic bracket/quote pairing
-}
